only create texture at start of program
